move directory for card carousel
